troquei TODOS os codigos para padronizar, ou deu mto bom ou fudeu tudo

+++ neuralNetwork/motor_V_Jou.py
@@ -133,7 +133,7 @@ for i in range(len(neurons)):
             print("============")
             print("")
 
-            neuron_per_layer = neurons[i]/layers[j]
+            neuron_per_layer = int(neurons[i]/layers[j])
 
             input_dim = len(train_data.columns.drop(target))
             model = RegressionModel(input_dim, 1, neuron_per_layer, layers[j])
